+++ internal/commands/kubernetes/create.go
@@ -125,6 +125,12 @@ func (c *createCommand) InitCommand() {
 			"taint=\"env=dev:NoSchedule\","+
 			"taint=\"env=dev2:NoSchedule\"`",
 	)
+	fs.StringArrayVar(
+		&c.params.ControlPlaneIPFilter,
+		"control-plane-allow-ip",
+		[]string{},
+		"Allow cluster control-plane to be accessed from an IP address or a network CIDR, multiple can be declared..",
+	)
 	config.AddToggleFlag(fs, &c.params.privateNodeGroups, "private-node-groups", false, "Do not assign public IPs to worker nodes. If set, the attached network should have a NAT gateway configured to provide internet access to the worker nodes.")
 	fs.StringVar(&c.params.Zone, "zone", "", namedargs.ZoneDescription("cluster"))
 	config.AddToggleFlag(fs, &c.params.wait, "wait", false, "Wait for cluster to be in running state before returning.")